backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
httpx
pydantic-settings
jose
python-multipart
fastapi_mail